CornerIndex to Position conversion implemented

+++ Assets/Scripts/TriangleMapLayout.cs
@@ -13,13 +13,13 @@ namespace Assets.Scripts
 		public readonly Vector2 MapCenter;
 		public readonly TriangleDimensions TriangleDimensions;
 		public readonly MapOrientation Orientation;
-		public readonly float RingCount;
+		public readonly int RingCount;
 
 		public TriangleMapLayout(
 			Vector2 mapCenter, 
 			TriangleDimensions triangleDimensions, 
-			MapOrientation orientation, 
-			float ringCount)
+			MapOrientation orientation,
+			int ringCount)
 		{
 			MapCenter = mapCenter;
 			TriangleDimensions = triangleDimensions;
